# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Decentra API",
        "description": "Description of your API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Login & Register"
                ],
                "summary": "Authenticate user and generate JWT token",
                "operationId": "2bd778ac594b753b1466d635caf32e28",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User's email Or UserName",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "User's password",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "password"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/v1/verify_wallet": {
            "get": {
                "tags": [
                    "Login & Register"
                ],
                "summary": "Validate user's wallet address for a specific network",
                "operationId": "a3a9d34a8e2ffe39fb6d73a8a63a5d74",
                "parameters": [
                    {
                        "name": "wallet_address",
                        "in": "query",
                        "description": "User's wallet address",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Address validation result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "valid": {
                                            "description": "Validation status",
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "description": "Validation message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}